more tuning updates

+++ trainer.py
@@ -7,10 +7,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from tester import test
-from testerperreloptimized import Testerperreloptimized
-from measureperrel import Measureperrel
 from dataload import DataLoader
-from testermine import test
 
 def train(dataloader, args, device='cuda'):
     # get model, dataset and optimizer
@@ -52,7 +49,7 @@ def train(dataloader, args, device='cuda'):
             kg_score_loss = args.kg_lambda * model.loss(kg_score, kg_x[:,3])
             reg_loss = args.reg_lambda * model.reg_loss()
 
-            loss = (rec_score_loss + kg_score_loss + reg_loss).to(torch.float32)
+            loss = rec_score_loss + kg_score_loss + reg_loss
 
             # step
             loss.backward()
